Renamed _maintenance_error to maintenance_error
